requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.6
django-debug-toolbar==4.4.2
django-environ==0.4.5
django-filter==24.2
djangorestframework==3.15.2
Faker==25.9.1
filelock==3.15.3
flake8==7.1.0
identify==2.5.36
idna==3.7
mccabe==0.7.0
nodeenv==1.9.1
platformdirs==4.2.2
pre-commit==3.7.1
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.2
virtualenv==20.26.3
